+++ test/register.js
@@ -1,0 +1,5 @@
+require('babel-register')({
+  ignore: /node_modules\/(?!@opuscapita\/react-perfect-scrollbar)/,
+});
+require('ignore-styles');
+require('global-jsdom')();